Guard-Interval VS BER
Script errorDistPlot now shows the distribution of the error along the samples of the switching cycle, and shows the resulting BER for the given guard-interval. Results are probably satisfactory.
The output signals on sw_filter now are contents of cells, since they can change size, specially when sampled. Other scripts are now adapted for that.
Got rid of the rls_i, since it has no relevance anymore.
Other changes are organization and code cleaning.